# backend/openapi/schema.yaml
openapi: 3.0.3
info:
  title: Project Meridian API
  version: v1
  description: Product lifecycle management platform API.
paths:
...
  /api/v1/auth/capabilities:
    get:
      operationId: auth_capabilities_retrieve
      tags:
      - auth
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthCapabilitiesResponse'
          description: ''
  /api/v1/auth/csrf:
    get:
      operationId: auth_csrf_retrieve
      tags:
      - auth
      security:
      - {}
      responses:
        '204':
          description: No response body
  /api/v1/auth/dev/login:
    post:
      operationId: auth_dev_login
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevLoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DevLoginRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DevLoginRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevLoginResponse'
          description: ''
        '404':
          description: No response body
  /api/v1/auth/dingtalk/callback:
    get:
      operationId: auth_dingtalk_callback
      summary: DingTalk OAuth callback
      tags:
      - auth
      security:
      - {}
      responses:
        '302':
          description: No response body
  /api/v1/auth/dingtalk/start:
    get:
      operationId: auth_dingtalk_start
      summary: Start DingTalk OAuth flow
      tags:
      - auth
      security:
      - {}
      responses:
        '302':
          description: No response body
  /api/v1/auth/logout:
    post:
      operationId: auth_logout
      tags:
      - auth
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/auth/pilot/login:
    post:
      operationId: auth_pilot_login
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PilotLoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PilotLoginRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PilotLoginRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PilotLoginResponse'
          description: ''
        '404':
          description: No response body
...
  /api/v1/documents/uploads:
    post:
      operationId: document_upload_sessions_create
      tags:
      - documents
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadSessionCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UploadSessionCreateRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSessionCreateResponse'
          description: ''
  /api/v1/documents/uploads/{public_id}/complete:
    post:
      operationId: document_upload_sessions_complete
      parameters:
      - in: path
        name: public_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - documents
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadSessionCompleteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UploadSessionCompleteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadSessionCompleteRequest'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSessionCompleteResponse'
          description: ''
...
components:
  schemas:
...
    AuthCapabilitiesResponse:
      type: object
      properties:
        pilot_password_login:
          type: boolean
        dev_login:
          type: boolean
      required:
      - dev_login
      - pilot_password_login
...
    DevLoginRequest:
      type: object
      properties:
        login_key:
          type: string
      required:
      - login_key
    DevLoginResponse:
      type: object
      properties:
        public_id:
          type: string
        display_name:
          type: string
      required:
      - display_name
      - public_id
...
    PilotLoginRequest:
      type: object
      properties:
        organization_public_id:
          type: string
          format: uuid
        employee_no:
          type: string
        password:
          type: string
      required:
      - employee_no
      - organization_public_id
      - password
    PilotLoginResponse:
      type: object
      properties:
        public_id:
          type: string
        display_name:
          type: string
      required:
      - display_name
      - public_id
...
    UploadSessionCompleteRequest:
      type: object
      properties:
        document_code:
          type: string
        title:
          type: string
    UploadSessionCompleteResponse:
      type: object
      properties:
        version_public_id:
          type: string
        document_public_id:
          type: string
        status:
          type: string
      required:
      - document_public_id
      - status
      - version_public_id
    UploadSessionCreateRequest:
      type: object
      properties:
        file:
          type: string
          format: uri
        original_filename:
          type: string
        declared_mime_type:
          type: string
        catalog_item_code:
          type: string
      required:
      - file
    UploadSessionCreateResponse:
      type: object
      properties:
        public_id:
          type: string
        original_filename:
          type: string
        declared_mime_type:
          type: string
        size_bytes:
          type: integer
        catalog_item_code:
          type: string
      required:
      - catalog_item_code
      - declared_mime_type
      - original_filename
      - public_id
      - size_bytes
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid